fix: wrong type

+++ data/src/main/java/com/jdevs/timeo/data/activities/ActivitiesDao.kt
@@ -23,5 +23,5 @@ interface ActivitiesDao : BaseDao<DBActivity> {
     fun getActivity(id: Int): Flow<DBActivity>
 
     @Query("UPDATE activities SET totalTime = totalTime + :by WHERE id = :id")
-    suspend fun increaseTime(id: Int, by: Int)
+    suspend fun increaseTime(id: Int, by: Long)
 }
